+++ konva.js
@@ -2,7 +2,7 @@
  * Konva JavaScript Framework v2.4.1
  * http://konvajs.github.io/
  * Licensed under the MIT
- * Date: Mon Oct 08 2018
+ * Date: Fri Oct 12 2018
  *
  * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
  * Modified work Copyright (C) 2014 - present by Anton Lavrenov (Konva)
@@ -3332,10 +3332,11 @@
      * @memberof Konva.Node.prototype
      * @returns {Boolean}
      */
-    shouldDrawHit: function(canvas) {
+    shouldDrawHit: function() {
       var layer = this.getLayer();
+
       return (
-        (canvas && canvas.isCache) ||
+        (!layer && this.isListening() && this.isVisible()) ||
         (layer &&
           layer.hitGraphEnabled() &&
           this.isListening() &&